BETA TEST

+++ app/user_dashboard/script_history/action/delete_action.ts
@@ -2,17 +2,18 @@
 import { prisma } from "@/lib/prisma"
 import { time } from "console"
 import { revalidatePath } from "next/cache"
-export async function delete_item(id:string) {
- 
+export async function delete_item(formData: FormData) {
+  const id = formData.get('idValue') as string;
   
-    const deletedItem = await prisma.script.delete({
-        where:{
-            Id:id
-        }
-    })
+  if (!id) {
+    throw new Error('ID manquant');
+  }
   
-   revalidatePath("/user_dashboard/script_history") 
-
-
-    
+  const deletedItem = await prisma.script.delete({
+    where: {
+      Id: id
+    }
+  });
+  
+  revalidatePath("/user_dashboard/script_history");
 }
